+++ src/app/services/global.service.ts
@@ -200,4 +200,23 @@ export class GlobalService {
   getCurrency(amount: any = 0) {
     return new Intl.NumberFormat('es-CO', { style: 'currency', currency: 'COP' }).format(amount);
   }
+
+  downloadCSVFromJson(filename, arrayOfJson) {
+    // convert JSON to CSV
+    const replacer = (key, value) => value === null ? '' : value; // specify how you want to handle null values here
+    const header = Object.keys(arrayOfJson[0]);
+    let csv = arrayOfJson.map(row => header.map(fieldName =>
+      JSON.stringify(row[fieldName], replacer)).join(','));
+    csv.unshift(header.join(','));
+    csv = csv.join('\r\n');
+
+    // Create link and download
+    const link = document.createElement('a');
+    link.setAttribute('href', 'data:text/csv;charset=utf-8,%EF%BB%BF' + encodeURIComponent(csv));
+    link.setAttribute('download', filename);
+    link.style.visibility = 'hidden';
+    document.body.appendChild(link);
+    link.click();
+    document.body.removeChild(link);
+  };
 }
