+++ src/app/views/grupos-familiares/reportes/reportes.page.ts
@@ -135,122 +135,6 @@ export class ReportesPage implements OnInit {
     else this.globalSer.presentAlert('Alerta', 'Disculpe, pero debe indicar una fecha inicial');
   }
 
-  async setDataToReport(data: any = null) {
-    await this.globalSer.presentLoading();
-    this.dataToReport = null;
-    this.dataToReport = !data ? null : {
-      data1: {
-        title: 'Hermanos',
-        data: [
-          {
-            label: 'Hermanos',
-            qty: data.brethren
-          },
-          {
-            label: 'Amigos',
-            qty: data.friends
-          }
-        ]
-      },
-      data2: {
-        title: 'Niños',
-        data: [
-          {
-            label: 'Niños',
-            qty: data.christianChildren
-          },
-          {
-            label: 'Amigos',
-            qty: data.christianChildrenFriends
-          },
-        ]
-      },
-      data3: {
-        title: 'Visitas',
-        data: [
-          {
-            label: 'Programadas',
-            qty: data.scheduledVisits
-          },
-          {
-            label: 'Disipulado',
-            qty: data.discipleshipVisits
-          },
-        ]
-      },
-      data4: {
-        title: 'Asistencia',
-        data: [
-          {
-            label: 'Hermanos',
-            qty: data.churchAttendance
-          },
-          {
-            label: 'Niños',
-            qty: data.churchAttendanceChildren
-          },
-        ]
-      },
-      data5: {
-        title: 'Conversiones y reconcilaciones',
-        data: [
-          {
-            label: 'Reconciliaciones',
-            qty: data.reconciliations
-          },
-          {
-            label: 'Conversiones',
-            qty: data.conversions
-          },
-          {
-            label: 'Conversiones de niños',
-            qty: data.conversionsChildren
-          },
-        ]
-      },
-      data6: {
-        title: 'Otros',
-        data: [
-          {
-            label: 'Planificación de hnos.',
-            qty: data.brethrenPlanning
-          },
-          {
-            label: 'Lectura bíblica',
-            qty: data.bibleReading
-          },
-          {
-            label: 'Consolidados',
-            qty: data.consolidated
-          },
-        ]
-      },
-      totals: {
-        title: 'Total de hermanos y niños',
-        qty: data.total
-      },
-      offering: {
-        title: 'Total de ofrendas',
-        qty: formatCurrency(data.offering, 'es-CO', 'COP')
-      },
-    };
-    await this.globalSer.dismissLoading();
-  }
-
-  async setObservationsValues(data: any) {
-    this.observations = [];
-    this.observationsPreview = [];
-    if (data && data.length > 0) {
-      this.observations = data;
-      data.forEach(o => {
-        this.observationsPreview.push({
-          observation: o.observations.length > 50 ? `${o.observations.substr(0, 50)} ...` : o.observations,
-          date: dayjs(o.date).format('DD-MM-YYYY HH:mm'),
-        });
-      });
-    }
-  }
-
   // actions inputs
   validateOnlyNumber(event: any) {
     onlyNumbersInputValidation2(event);
