+++ src/app/views/miembros/miembros.page.html
@@ -15,7 +15,7 @@
     <ion-row>
       <ion-col size="12">
         <ion-row>
-          <ion-col [size]="12" size-sm="7">
+          <ion-col [size]="12" size-sm="6">
             <ion-toolbar>
               <ion-button color="medium" (click)="queryParams.word = null; findData()" slot="start" *ngIf="queryParams.word">
                 <ion-icon name="backspace-outline"></ion-icon>
@@ -33,7 +33,7 @@
               </ion-button>
             </ion-toolbar>
           </ion-col>
-          <ion-col [size]="12" size-sm="5">
+          <ion-col [size]="12" size-sm="6">
             <ion-toolbar>
               <ion-button color="medium" (click)="setSortOrder()">
                 <!--<ion-icon name="funnel-outline" slot="start"></ion-icon>-->
@@ -44,6 +44,10 @@
               <ion-button color="medium" (click)="setQueryValues()">
                 <ion-icon name="filter-outline"></ion-icon>&nbsp;
                 <ion-text>{{ queryParams.limit }}</ion-text>
+              </ion-button>
+              <ion-button (click)="downloadData()">
+                <ion-icon name="download-outline" slot="start"></ion-icon>
+                <ion-text>Descargar datos</ion-text>
               </ion-button>
             </ion-toolbar>
           </ion-col>
