--- www/default~detalles-miembro-detalles-miembro-module~registro-registro-module~views-mi-cuenta-mi-cuenta-module.js
@@ -1,1597 +1,0 @@
-(window["webpackJsonp"] = window["webpackJsonp"] || []).push([["default~detalles-miembro-detalles-miembro-module~registro-registro-module~views-mi-cuenta-mi-cuenta-module"],{
-
-/***/ "0ox5":
-/*!***********************************!*\
-  !*** ./src/Utils/profile.data.ts ***!
-  \***********************************/
-/*! exports provided: educationLevels, bloodType, professions, companyType, civilStatus, gender, documentType */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "educationLevels", function() { return educationLevels; });
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "bloodType", function() { return bloodType; });
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "professions", function() { return professions; });
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "companyType", function() { return companyType; });
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "civilStatus", function() { return civilStatus; });
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "gender", function() { return gender; });
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "documentType", function() { return documentType; });
-const educationLevels = [
-    'Preescolar',
-    'Básica Primaria (1-5)',
-    'Básica Secundaria (6-9)',
-    'Média (10-13)',
-    'Técnica Laboral',
-    'Técnica Profesional',
-    'Tecnológica',
-    'Universitaria',
-    'Especialización',
-    'Maestría',
-    'Doctorado'
-];
-const bloodType = ['A -', 'A +', 'AB -', 'AB +', 'B -', 'B +', 'O -', 'O +'];
-const professions = [
-    'Prefiero no indicar',
-    'Abogado',
-    'Actor, Actriz, Artista, Director de Espectáculos',
-    'Actuario',
-    'Aduanero/Agente de Aduanas/Inspectores de Frontera',
-    'Aeromozo/Azafata',
-    'Agente/Intermediario/Corredor Inmobiliario',
-    'Agente de Bolsa',
-    'Agente de Inmigración/Migración',
-    'Agente de Turismo/Viajes',
-    'Agente/Intermediario/Corredor de Seguros',
-    'Agricultor, Agrónomo, Agrologo, Arboricultor',
-    'Administrador',
-    'Albañil, Obrero de Construcción',
-    'Ama de Casa',
-    'Analista de Sistema y Computación',
-    'Antropólogo, Arqueólogo y Etnólogo',
-    'Archivero',
-    'Armador de Barco',
-    'Arquitecto Artesano',
-    'Asistente Social',
-    'Autor Literario, Escritor y Crmtico',
-    'Bacteriólogo, Farmacólogo, Biólogo, Científico',
-    'Basurero/Barrendero',
-    'Cajero',
-    'Camarero/Barman/Mesero/Chef',
-    'Cambista, Compra/Venta de Moneda',
-    'Campesino',
-    'Capataz',
-    'Cargador',
-    'Carpintero',
-    'Cartero',
-    'Cerrajero',
-    'Cobrador',
-    'Comerciante / Vendedor',
-    'Conductor, Chofer / Taxista',
-    'Conserje/Portero/Guardián/Vigilante',
-    'Constructor',
-    'Contador',
-    'Contratista',
-    'Corte y Confección de Ropa/Fabricante de Prendas',
-    'Cosmetólogo, Peluquero y Barbero',
-    'Decorador, Dibujante, Publicista',
-    'Dentista / Odontólogo',
-    'Deportista Profesional, Atleta, Arbitro',
-    'Distribuidor',
-    'Docente',
-    'Economista',
-    'Electricista',
-    'Empleada(o) del hogar / Nana',
-    'Empresario Exportador/Empresario Importador',
-    'Enfermero',
-    'Ensamblador',
-    'Escultor',
-    'Estudiante',
-    'Fotógrafo/Operadores cámara, cine y tv, locutor',
-    'Ganadero',
-    'Gasfitero',
-    'Historiador',
-    'Ingeniero',
-    'Interprete, Traductor',
-    'Jardinero',
-    'Jockey',
-    'Joyero y/o Platero / Orfebre',
-    'Jubilado / Pensionista',
-    'Laboratorista (Técnico)',
-    'Liquidador, Reclamaciones/Seguros',
-    'Maquinista / Operador de maquinaria',
-    'Martillero / Subastador',
-    'Mayorista, comercio al por mayor',
-    'Mecánico',
-    'Medico / Cirujano',
-    'Metalurgista',
-    'Miembro de las Fuerzas Armadas',
-    'Nutricionista',
-    'Obrero / Operador',
-    'Obstetriz',
-    'Organizador de Eventos',
-    'Panadero / Pastelero',
-    'Paramédico',
-    'Periodista',
-    'Perito',
-    'Pescador',
-    'Piloto',
-    'Pintor',
-    'Policía Municipal',
-    'Policía PNP',
-    'Productor de Cine / Radio / Televisión / Teatro',
-    'Productor, Cultivos Extensivos',
-    'Programador',
-    'Psicólogo / Terapeuta',
-    'Quiropráctico/Kinesiterapeuta (Kinesiólogos)',
-    'Relacionista Publico e Industrial',
-    'Relojero',
-    'Reparación de Automóviles, Pintor Retocador',
-    'Reparador de Aparatos Electrodomésticos',
-    'Repartidor',
-    'Sacerdote/Monja',
-    'Secretaria, Recepcionista, Telefonista',
-    'Seguridad/Guardaespaldas/Guardia de Seguridad',
-    'Servicio de Almacenamiento/Almacenero',
-    'Servicio de Alquiler de Vehículos',
-    'Servicios de Alquiler de Videos, Equipos de Sonido',
-    'Sociólogo',
-    'Tasador',
-    'Técnico',
-    'Torero',
-    'Tramitador',
-    'Transporte de Carga y/o Mudanza',
-    'Transportista',
-    'Vendedor Ambulante',
-    'Veterinario, Zoólogo, Zootécnico',
-    'Visitador Medico',
-    'Zapatero',
-    'Otros',
-];
-const companyType = [
-    'Empresa asociativa de trabajo',
-    'Empresa unipersonal',
-    'Persona natural comerciante',
-    'Sociedad anónima',
-    'Sociedad colectiva',
-    'Sociedad en comandita por acciones',
-    'Sociedad en comandita simple',
-    'Sociedad limitada',
-    'Sociedades agrarias de transformación',
-    'Sociedades por acciones simplificadas',
-    'Otro',
-];
-const civilStatus = [
-    'Soltero',
-    'Casado',
-    'Divorciado',
-    'Separado',
-    'Prefiero no indicar',
-];
-const gender = [
-    'Hombre',
-    'Mujer',
-    'Prefiero no indicar',
-];
-const documentType = [
-    { val: 'CC', label: 'Cédula de ciudadanía' },
-    { val: 'TI', label: 'Tarjeta de identidad' },
-    { val: 'PAS', label: 'Pasaporte' },
-    { val: 'CE', label: 'Cédula de extrajería' },
-    { val: 'PE', label: 'Permiso Especial de Permanencia' },
-];
-
-
-/***/ }),
-
-/***/ "ZrT4":
-/*!*************************************!*\
-  !*** ./src/Utils/locations.data.ts ***!
-  \*************************************/
-/*! exports provided: departments */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
-
-"use strict";
-__webpack_require__.r(__webpack_exports__);
-/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "departments", function() { return departments; });
-const departments = [
-    {
-        department: 'Amazonas',
-        cities: [
-            'Leticia'
-        ]
-    },
-    {
-        department: 'Antioquia',
-        cities: [
-            'Abejorral',
-            'Abriaquí',
-            'Alejandría',
-            'Amagá',
-            'Amalfi',
-            'Andes',
-            'Angelópolis',
-            'Angostura',
-            'Anorí',
-            'Anzá',
-            'Apartadó',
-            'Arboletes',
-            'Argelia',
-            'Armenia',
-            'Barbosa',
-            'Bello',
-            'Belmira',
-            'Betania',
-            'Betulia',
-            'Briceño',
-            'Buriticá',
-            'Cáceres',
-            'Caicedo',
-            'Caldas',
-            'Campamento',
-            'Cañasgordas',
-            'Caracolí',
-            'Caramanta',
-            'Carepa',
-            'Carolina del Príncipe',
-            'Caucasia',
-            'Chigorodó',
-            'Cisneros',
-            'Ciudad Bolívar',
-            'Cocorná',
-            'Concepción',
-            'Concordia',
-            'Copacabana',
-            'Dabeiba',
-            'Donmatías',
-            'Ebéjico',
-            'El Bagre',
-            'El Carmen de Viboral',
-            'El Peñol',
-            'El Retiro',
-            'El Santuario',
-            'Entrerríos',
-            'Envigado',
-            'Fredonia',
-            'Frontino',
-            'Giraldo',
-            'Girardota',
-            'Gómez Plata',
-            'Granada',
-            'Guadalupe',
-            'Guarne',
-            'Guatapé',
-            'Heliconia',
-            'Hispania',
-            'Itagüí',
-            'Ituango',
-            'Jardín',
-            'Jericó',
-            'La Ceja',
-            'La Estrella',
-            'La Pintada',
-            'La Unión',
-            'Liborina',
-            'Maceo',
-            'Marinilla',
-            'Medellín',
-            'Montebello',
-            'Murindó',
-            'Mutatá',
-            'Nariño',
-            'Nechí',
-            'Necoclí',
-            'Olaya',
-            'Peque',
-            'Pueblorrico',
-            'Puerto Berrío',
-            'Puerto Nare',
-            'Puerto Triunfo',
-            'Remedios',
-            'Rionegro',
-            'Sabanalarga',
-            'Sabaneta',
-            'Salgar',
-            'San Andrés de Cuerquia',
-            'San Carlos',
-            'San Francisco',
-            'San Jerónimo',
-            'San José de la Montaña',
-            'San Juan de Urabá',
-            'San Luis',
-            'San Pedro de Urabá',
-            'San Pedro de los Milagros',
-            'San Rafael',
-            'San Roque',
-            'San Vicente',
-            'Santa Bárbara',
-            'Santa Fe de Antioquia',
-            'Santa Rosa de Osos',
-            'Santo Domingo',
-            'Segovia',
-            'Sonsón',
-            'Sopetrán',
-            'Támesis',
-            'Tarazá',
-            'Tarso',
-            'Titiribí',
-            'Toledo',
-            'Turbo',
-            'Uramita',
-            'Urrao',
-            'Valdivia',
-            'Valparaíso',
-            'Vegachí',
-            'Venecia',
-            'Vigía del Fuerte',
-            'Yalí',
-            'Yarumal',
-            'Yolombó',
-            'Yondó',
-            'Zaragoza'
-        ]
-    },
-    {
-        department: 'Arauca',
-        cities: [
-            'Arauca',
-            'Arauquita',
-            'Cravo Norte',
-            'Fortul',
-            'Puerto Rondón',
-            'Saravena',
-            'Tame'
-        ]
-    },
-    {
-        department: 'Atlántico',
-        cities: [
-            'Baranoa',
-            'Barranquilla',
-            'Campo de la Cruz',
-            'Candelaria',
-            'Galapa',
-            'Juan de Acosta',
-            'Luruaco',
-            'Malambo',
-            'Manatí',
-            'Palmar de Varela',
-            'Piojó',
-            'Polonuevo',
-            'Ponedera',
-            'Puerto Colombia',
-            'Repelón',
-            'Sabanagrande',
-            'Sabanalarga',
-            'Santa Lucía',
-            'Santo Tomás',
-            'Soledad',
-            'Suán',
-            'Tubará',
-            'Usiacurí'
-        ]
-    },
-    {
-        department: 'Bolívar',
-        cities: [
-            'Achí',
-            'Altos del Rosario',
-            'Arenal',
-            'Arjona',
-            'Arroyohondo',
-            'Barranco de Loba',
-            'Brazuelo de Papayal',
-            'Calamar',
-            'Cantagallo',
-            'Cartagena de Indias',
-            'Cicuco',
-            'Clemencia',
-            'Córdoba',
-            'El Carmen de Bolívar',
-            'El Guamo',
-            'El Peñón',
-            'Hatillo de Loba',
-            'Magangué',
-            'Mahates',
-            'Margarita',
-            'María la Baja',
-            'Mompós',
-            'Montecristo',
-            'Morales',
-            'Norosí',
-            'Pinillos',
-            'Regidor',
-            'Río Viejo',
-            'San Cristóbal',
-            'San Estanislao',
-            'San Fernando',
-            'San Jacinto del Cauca',
-            'San Jacinto',
-            'San Juan Nepomuceno',
-            'San Martín de Loba',
-            'San Pablo',
-            'Santa Catalina',
-            'Santa Rosa',
-            'Santa Rosa del Sur',
-            'Simití',
-            'Soplaviento',
-            'Talaigua Nuevo',
-            'Tiquisio',
-            'Turbaco',
-            'Turbaná',
-            'Villanueva',
-            'Zambrano'
-        ]
-    },
-    {
-        department: 'Boyacá',
-        cities: [
-            'Almeida',
-            'Aquitania',
-            'Arcabuco',
-            'Belén',
-            'Berbeo',
-            'Betéitiva',
-            'Boavita',
-            'Boyacá',
-            'Briceño',
-            'Buenavista',
-            'Busbanzá',
-            'Caldas',
-            'Campohermoso',
-            'Cerinza',
-            'Chinavita',
-            'Chiquinquirá',
-            'Chíquiza',
-            'Chiscas',
-            'Chita',
-            'Chitaraque',
-            'Chivatá',
-            'Chivor',
-            'Ciénega',
-            'Cómbita',
-            'Coper',
-            'Corrales',
-            'Covarachía',
-            'Cubará',
-            'Cucaita',
-            'Cuítiva',
-            'Duitama',
-            'El Cocuy',
-            'El Espino',
-            'Firavitoba',
-            'Floresta',
-            'Gachantivá',
-            'Gámeza',
-            'Garagoa',
-            'Guacamayas',
-            'Guateque',
-            'Guayatá',
-            'Güicán',
-            'Iza',
-            'Jenesano',
-            'Jericó',
-            'La Capilla',
-            'La Uvita',
-            'La Victoria',
-            'Labranzagrande',
-            'Macanal',
-            'Maripí',
-            'Miraflores',
-            'Mongua',
-            'Monguí',
-            'Moniquirá',
-            'Motavita',
-            'Muzo',
-            'Nobsa',
-            'Nuevo Colón',
-            'Oicatá',
-            'Otanche',
-            'Pachavita',
-            'Páez',
-            'Paipa',
-            'Pajarito',
-            'Panqueba',
-            'Pauna',
-            'Paya',
-            'Paz del Río',
-            'Pesca',
-            'Pisba',
-            'Puerto Boyacá',
-            'Quípama',
-            'Ramiriquí',
-            'Ráquira',
-            'Rondón',
-            'Saboyá',
-            'Sáchica',
-            'Samacá',
-            'San Eduardo',
-            'San José de Pare',
-            'San Luis de Gaceno',
-            'San Mateo',
-            'San Miguel de Sema',
-            'San Pablo de Borbur',
-            'Santa María',
-            'Santa Rosa de Viterbo',
-            'Santa Sofía',
-            'Santana',
-            'Sativanorte',
-            'Sativasur',
-            'Siachoque',
-            'Soatá',
-            'Socha',
-            'Socotá',
-            'Sogamoso',
-            'Somondoco',
-            'Sora',
-            'Soracá',
-            'Sotaquirá',
-            'Susacón',
-            'Sutamarchán',
-            'Sutatenza',
-            'Tasco',
-            'Tenza',
-            'Tibaná',
-            'Tibasosa',
-            'Tinjacá',
-            'Tipacoque',
-            'Toca',
-            'Togüí',
-            'Tópaga',
-            'Tota',
-            'Tunja',
-            'Tununguá',
-            'Turmequé',
-            'Tuta',
-            'Tutazá',
-            'Úmbita',
-            'Ventaquemada',
-            'Villa de Leyva',
-            'Viracachá',
-            'Zetaquira'
-        ]
-    },
-    {
-        department: 'Caldas',
-        cities: [
-            'Aguadas',
-            'Anserma',
-            'Aranzazu',
-            'Belalcázar',
-            'Chinchiná',
-            'Filadelfia',
-            'La Dorada',
-            'La Merced',
-            'Manizales',
-            'Manzanares',
-            'Marmato',
-            'Marquetalia',
-            'Marulanda',
-            'Neira',
-            'Norcasia',
-            'Pácora',
-            'Palestina',
-            'Pensilvania',
-            'Riosucio',
-            'Risaralda',
-            'Salamina',
-            'Samaná',
-            'San José',
-            'Supía',
-            'Victoria',
-            'Villamaría',
-            'Viterbo'
-        ]
-    },
-    {
-        department: 'Caquetá',
-        cities: [
-            'Albania',
-            'Belén de los Andaquíes',
-            'Cartagena del Chairá',
-            'Curillo',
-            'El Doncello',
-            'El Paujil',
-            'Florencia',
-            'La Montañita',
-            'Milán',
-            'Morelia',
-            'Puerto Rico',
-            'San José del Fragua',
-            'San Vicente del Caguán',
-            'Solano',
-            'Solita',
-            'Valparaíso'
-        ]
-    },
-    {
-        department: 'Casanare',
-        cities: [
-            'Aguazul',
-            'Chámeza',
-            'Hato Corozal',
-            'La Salina',
-            'Maní',
-            'Monterrey',
-            'Nunchía',
-            'Orocué',
-            'Paz de Ariporo',
-            'Pore',
-            'Recetor',
-            'Sabanalarga',
-            'Sácama',
-            'San Luis de Palenque',
-            'Támara',
-            'Tauramena',
-            'Trinidad',
-            'Villanueva',
-            'Yopal'
-        ]
-    },
-    {
-        department: 'Cauca',
-        cities: [
-            'Almaguer',
-            'Argelia',
-            'Balboa',
-            'Bolívar',
-            'Buenos Aires',
-            'Cajibío',
-            'Caldono',
-            'Caloto',
-            'Corinto',
-            'El Tambo',
-            'Florencia',
-            'Guachené',
-            'Guapí',
-            'Inzá',
-            'Jambaló',
-            'La Sierra',
-            'La Vega',
-            'López de Micay',
-            'Mercaderes',
-            'Miranda',
-            'Morales',
-            'Padilla',
-            'Páez',
-            'Patía',
-            'Piamonte',
-            'Piendamó',
-            'Popayán',
-            'Puerto Tejada',
-            'Puracé',
-            'Rosas',
-            'San Sebastián',
-            'Santa Rosa',
-            'Santander de Quilichao',
-            'Silvia',
-            'Sotará',
-            'Suárez',
-            'Sucre',
-            'Timbío',
-            'Timbiquí',
-            'Toribío',
-            'Totoró',
-            'Villa Rica'
-        ]
-    },
-    {
-        department: 'Cesar',
-        cities: [
-            'Aguachica',
-            'Agustín Codazzi',
-            'Astrea',
-            'Becerril',
-            'Bosconia',
-            'Chimichagua',
-            'Chiriguaná',
-            'Curumaní',
-            'El Copey',
-            'El Paso',
-            'Gamarra',
-            'González',
-            'La Gloria (Cesar)',
-            'La Jagua de Ibirico',
-            'La Paz',
-            'Manaure Balcón del Cesar',
-            'Pailitas',
-            'Pelaya',
-            'Pueblo Bello',
-            'Río de Oro',
-            'San Alberto',
-            'San Diego',
-            'San Martín',
-            'Tamalameque',
-            'Valledupar'
-        ]
-    },
-    {
-        department: 'Chocó',
-        cities: [
-            'Acandí',
-            'Alto Baudó',
-            'Bagadó',
-            'Bahía Solano',
-            'Bajo Baudó',
-            'Bojayá',
-            'Cantón de San Pablo',
-            'Cértegui',
-            'Condoto',
-            'El Atrato',
-            'El Carmen de Atrato',
-            'El Carmen del Darién',
-            'Istmina',
-            'Juradó',
-            'Litoral de San Juan',
-            'Lloró',
-            'Medio Atrato',
-            'Medio Baudó',
-            'Medio San Juan',
-            'Nóvita',
-            'Nuquí',
-            'Quibdó',
-            'Río Iró',
-            'Río Quito',
-            'Riosucio',
-            'San José del Palmar',
-            'Sipí',
-            'Tadó',
-            'Unión Panamericana',
-            'Unguía'
-        ]
-    },
-    {
-        department: 'Cundinamarca',
-        cities: [
-            'Agua de Dios',
-            'Albán',
-            'Anapoima',
-            'Anolaima',
-            'Apulo',
-            'Arbeláez',
-            'Beltrán',
-            'Bituima',
-            'Bogotá',
-            'Bojacá',
-            'Cabrera',
-            'Cachipay',
-            'Cajicá',
-            'Caparrapí',
-            'Cáqueza',
-            'Carmen de Carupa',
-            'Chaguaní',
-            'Chía',
-            'Chipaque',
-            'Choachí',
-            'Chocontá',
-            'Cogua',
-            'Cota',
-            'Cucunubá',
-            'El Colegio',
-            'El Peñón',
-            'El Rosal',
-            'Facatativá',
-            'Fómeque',
-            'Fosca',
-            'Funza',
-            'Fúquene',
-            'Fusagasugá',
-            'Gachalá',
-            'Gachancipá',
-            'Gachetá',
-            'Gama',
-            'Girardot',
-            'Granada',
-            'Guachetá',
-            'Guaduas',
-            'Guasca',
-            'Guataquí',
-            'Guatavita',
-            'Guayabal de Síquima',
-            'Guayabetal',
-            'Gutiérrez',
-            'Jerusalén',
-            'Junín',
-            'La Calera',
-            'La Mesa',
-            'La Palma',
-            'La Peña',
-            'La Vega',
-            'Lenguazaque',
-            'Machetá',
-            'Madrid',
-            'Manta',
-            'Medina',
-            'Mosquera',
-            'Nariño',
-            'Nemocón',
-            'Nilo',
-            'Nimaima',
-            'Nocaima',
-            'Pacho',
-            'Paime',
-            'Pandi',
-            'Paratebueno',
-            'Pasca',
-            'Puerto Salgar',
-            'Pulí',
-            'Quebradanegra',
-            'Quetame',
-            'Quipile',
-            'Ricaurte',
-            'San Antonio del Tequendama',
-            'San Bernardo',
-            'San Cayetano',
-            'San Francisco',
-            'San Juan de Rioseco',
-            'Sasaima',
-            'Sesquilé',
-            'Sibaté',
-            'Silvania',
-            'Simijaca',
-            'Soacha',
-            'Sopó',
-            'Subachoque',
-            'Suesca',
-            'Supatá',
-            'Susa',
-            'Sutatausa',
-            'Tabio',
-            'Tausa',
-            'Tena',
-            'Tenjo',
-            'Tibacuy',
-            'Tibirita',
-            'Tocaima',
-            'Tocancipá',
-            'Topaipí',
-            'Ubalá',
-            'Ubaque',
-            'Ubaté',
-            'Une',
-            'Útica',
-            'Venecia',
-            'Vergara',
-            'Vianí',
-            'Villagómez',
-            'Villapinzón',
-            'Villeta',
-            'Viotá',
-            'Yacopí',
-            'Zipacón',
-            'Zipaquirá'
-        ]
-    },
-    {
-        department: 'Córdoba',
-        cities: [
-            'Ayapel',
-            'Buenavista',
-            'Canalete',
-            'Cereté',
-            'Chimá',
-            'Chinú',
-            'Ciénaga de Oro',
-            'Cotorra',
-            'La Apartada',
-            'Lorica',
-            'Los Córdobas',
-            'Momil',
-            'Montelíbano',
-            'Montería',
-            'Moñitos',
-            'Planeta Rica',
-            'Pueblo Nuevo',
-            'Puerto Escondido',
-            'Puerto Libertador',
-            'Purísima',
-            'Sahagún',
-            'San Andrés de Sotavento',
-            'San Antero',
-            'San Bernardo del Viento',
-            'San Carlos',
-            'San José de Uré',
-            'San Pelayo',
-            'Tierralta',
-            'Tuchín',
-            'Valencia'
-        ]
-    },
-    {
-        department: 'Guainía',
-        cities: [
-            'Inírida'
-        ]
-    },
-    {
-        department: 'Guaviare',
-        cities: [
-            'Calamar',
-            'El Retorno',
-            'Miraflores',
-            'San José del Guaviare'
-        ]
-    },
-    {
-        department: 'Huila',
-        cities: [
-            'Acevedo',
-            'Agrado',
-            'Aipe',
-            'Algeciras',
-            'Altamira',
-            'Baraya',
-            'Campoalegre',
-            'Colombia',
-            'El Pital',
-            'Elías',
-            'Garzón',
-            'Gigante',
-            'Guadalupe',
-            'Hobo',
-            'Íquira',
-            'Isnos',
-            'La Argentina',
-            'La Plata',
-            'Nátaga',
-            'Neiva',
-            'Oporapa',
-            'Paicol',
-            'Palermo',
-            'Palestina',
-            'Pitalito',
-            'Rivera',
-            'Saladoblanco',
-            'San Agustín',
-            'Santa María',
-            'Suaza',
-            'Tarqui',
-            'Tello',
-            'Teruel',
-            'Tesalia',
-            'Timaná',
-            'Villavieja',
-            'Yaguará'
-        ]
-    },
-    {
-        department: 'La Guajira',
-        cities: [
-            'Albania',
-            'Barrancas',
-            'Dibulla',
-            'Distracción',
-            'El Molino',
-            'Fonseca',
-            'Hatonuevo',
-            'La Jagua del Pilar',
-            'Maicao',
-            'Manaure',
-            'Riohacha',
-            'San Juan del Cesar',
-            'Uribia',
-            'Urumita',
-            'Villanueva'
-        ]
-    },
-    {
-        department: 'Magdalena',
-        cities: [
-            'Algarrobo',
-            'Aracataca',
-            'Ariguaní',
-            'Cerro de San Antonio',
-            'Chibolo',
-            'Chibolo',
-            'Ciénaga',
-            'Concordia',
-            'El Banco',
-            'El Piñón',
-            'El Retén',
-            'Fundación',
-            'Guamal',
-            'Nueva Granada',
-            'Pedraza',
-            'Pijiño del Carmen',
-            'Pivijay',
-            'Plato',
-            'Pueblo Viejo',
-            'Remolino',
-            'Sabanas de San Ángel',
-            'Salamina',
-            'San Sebastián de Buenavista',
-            'San Zenón',
-            'Santa Ana',
-            'Santa Bárbara de Pinto',
-            'Santa Marta',
-            'Sitionuevo',
-            'Tenerife',
-            'Zapayán',
-            'Zona Bananera'
-        ]
-    },
-    {
-        department: 'Meta',
-        cities: [
-            'Acacías',
-            'Barranca de Upía',
-            'Cabuyaro',
-            'Castilla la Nueva',
-            'Cubarral',
-            'Cumaral',
-            'El Calvario',
-            'El Castillo',
-            'El Dorado',
-            'Fuente de Oro',
-            'Granada',
-            'Guamal',
-            'La Macarena',
-            'La Uribe',
-            'Lejanías',
-            'Mapiripán',
-            'Mesetas',
-            'Puerto Concordia',
-            'Puerto Gaitán',
-            'Puerto Lleras',
-            'Puerto López',
-            'Puerto Rico',
-            'Restrepo',
-            'San Carlos de Guaroa',
-            'San Juan de Arama',
-            'San Juanito',
-            'San Martín',
-            'Villavicencio',
-            'Vista Hermosa'
-        ]
-    },
-    {
-        department: 'Nariño',
-        cities: [
-            'Aldana',
-            'Ancuyá',
-            'Arboleda',
-            'Barbacoas',
-            'Belén',
-            'Buesaco',
-            'Chachagüí',
-            'Colón',
-            'Consacá',
-            'Contadero',
-            'Córdoba',
-            'Cuaspud',
-            'Cumbal',
-            'Cumbitara',
-            'El Charco',
-            'El Peñol',
-            'El Rosario',
-            'El Tablón',
-            'El Tambo',
-            'Francisco Pizarro',
-            'Funes',
-            'Guachucal',
-            'Guaitarilla',
-            'Gualmatán',
-            'Iles',
-            'Imués',
-            'Ipiales',
-            'La Cruz',
-            'La Florida',
-            'La Llanada',
-            'La Tola',
-            'La Unión',
-            'Leiva',
-            'Linares',
-            'Los Andes',
-            'Magüí Payán',
-            'Mallama',
-            'Mosquera',
-            'Nariño',
-            'Olaya Herrera',
-            'Ospina',
-            'Pasto',
-            'Policarpa',
-            'Potosí',
-            'Providencia',
-            'Puerres',
-            'Pupiales',
-            'Ricaurte',
-            'Roberto Payán',
-            'Samaniego',
-            'San Bernardo',
-            'San José de Albán',
-            'San Lorenzo',
-            'San Pablo',
-            'San Pedro de Cartago',
-            'Sandoná',
-            'Santa Bárbara',
-            'Santacruz',
-            'Sapuyes',
-            'Taminango',
-            'Tangua',
-            'Tumaco',
-            'Túquerres',
-            'Yacuanquer'
-        ]
-    },
-    {
-        department: 'Norte de Santander',
-        cities: [
-            'Ábrego',
-            'Arboledas',
-            'Bochalema',
-            'Bucarasica',
-            'Cáchira',
-            'Cácota',
-            'Chinácota',
-            'Chitagá',
-            'Convención',
-            'Cúcuta',
-            'Cucutilla',
-            'Duranía',
-            'El Carmen',
-            'El Tarra',
-            'El Zulia',
-            'Gramalote',
-            'Hacarí',
-            'Herrán',
-            'La Esperanza',
-            'La Playa de Belén',
-            'Labateca',
-            'Los Patios',
-            'Lourdes',
-            'Mutiscua',
-            'Ocaña',
-            'Pamplona',
-            'Pamplonita',
-            'Puerto Santander',
-            'Ragonvalia',
-            'Salazar de Las Palmas',
-            'San Calixto',
-            'San Cayetano',
-            'Santiago',
-            'Santo Domingo de Silos',
-            'Sardinata',
-            'Teorama',
-            'Tibú',
-            'Toledo',
-            'Villa Caro',
-            'Villa del Rosario'
-        ]
-    },
-    {
-        department: 'Putumayo',
-        cities: [
-            'Colón',
-            'Mocoa',
-            'Orito',
-            'Puerto Asís',
-            'Puerto Caicedo',
-            'Puerto Guzmán',
-            'Puerto Leguízamo',
-            'San Francisco',
-            'San Miguel',
-            'Santiago',
-            'Sibundoy',
-            'Valle del Guamuez',
-            'Villagarzón'
-        ]
-    },
-    {
-        department: 'Quindío',
-        cities: [
-            'Armenia',
-            'Buenavista',
-            'Calarcá',
-            'Circasia',
-            'Córdoba',
-            'Filandia',
-            'Génova',
-            'La Tebaida',
-            'Montenegro',
-            'Pijao',
-            'Quimbaya',
-            'Salento'
-        ]
-    },
-    {
-        department: 'Risaralda',
-        cities: [
-            'Apía',
-            'Balboa',
-            'Belén de Umbría',
-            'Dosquebradas',
-            'Guática',
-            'La Celia',
-            'La Virginia',
-            'Marsella',
-            'Mistrató',
-            'Pereira',
-            'Pueblo Rico',
-            'Quinchía',
-            'Santa Rosa de Cabal',
-            'Santuario'
-        ]
-    },
-    {
-        department: 'San Andrés y Providencia',
-        cities: [
-            'Providencia y Santa Catalina Islas',
-            'San Andrés'
-        ]
-    },
-    {
-        department: 'Santander',
-        cities: [
-            'Aguada',
-            'Albania',
-            'Aratoca',
-            'Barbosa',
-            'Barichara',
-            'Barrancabermeja',
-            'Betulia',
-            'Bolívar',
-            'Bucaramanga',
-            'Cabrera',
-            'California',
-            'Capitanejo',
-            'Carcasí',
-            'Cepitá',
-            'Cerrito',
-            'Charalá',
-            'Charta',
-            'Chima',
-            'Chipatá',
-            'Cimitarra',
-            'Concepción',
-            'Confines',
-            'Contratación',
-            'Coromoro',
-            'Curití',
-            'El Carmen de Chucurí',
-            'El Guacamayo',
-            'El Peñón',
-            'El Playón',
-            'El Socorro',
-            'Encino',
-            'Enciso',
-            'Florián',
-            'Floridablanca',
-            'Galán',
-            'Gámbita',
-            'Girón',
-            'Guaca',
-            'Guadalupe',
-            'Guapotá',
-            'Guavatá',
-            'Güepsa',
-            'Hato',
-            'Jesús María',
-            'Jordán',
-            'La Belleza',
-            'La Paz',
-            'Landázuri',
-            'Lebrija',
-            'Los Santos',
-            'Macaravita',
-            'Málaga',
-            'Matanza',
-            'Mogotes',
-            'Molagavita',
-            'Ocamonte',
-            'Oiba',
-            'Onzaga',
-            'Palmar',
-            'Palmas del Socorro',
-            'Páramo',
-            'Piedecuesta',
-            'Pinchote',
-            'Puente Nacional',
-            'Puerto Parra',
-            'Puerto Wilches',
-            'Rionegro',
-            'Sabana de Torres',
-            'San Andrés',
-            'San Benito',
-            'San Gil',
-            'San Joaquín',
-            'San José de Miranda',
-            'San Miguel',
-            'San Vicente de Chucurí',
-            'Santa Bárbara',
-            'Santa Helena del Opón',
-            'Simacota',
-            'Suaita',
-            'Sucre',
-            'Suratá',
-            'Tona',
-            'Valle de San José',
-            'Vélez',
-            'Vetas',
-            'Villanueva',
-            'Zapatoca'
-        ]
-    },
-    {
-        department: 'Sucre',
-        cities: [
-            'Buenavista',
-            'Caimito',
-            'Chalán',
-            'Colosó',
-            'Corozal',
-            'Coveñas',
-            'El Roble',
-            'Galeras',
-            'Guaranda',
-            'La Unión',
-            'Los Palmitos',
-            'Majagual',
-            'Morroa',
-            'Ovejas',
-            'Sampués',
-            'San Antonio de Palmito',
-            'San Benito Abad',
-            'San Juan de Betulia',
-            'San Marcos',
-            'San Onofre',
-            'San Pedro',
-            'Sincé',
-            'Sincelejo',
-            'Sucre',
-            'Tolú',
-            'Tolú Viejo'
-        ]
-    },
-    {
-        department: 'Tolima',
-        cities: [
-            'Alpujarra',
-            'Alvarado',
-            'Ambalema',
-            'Anzoátegui',
-            'Armero',
-            'Ataco',
-            'Cajamarca',
-            'Carmen de Apicalá',
-            'Casabianca',
-            'Chaparral',
-            'Coello',
-            'Coyaima',
-            'Cunday',
-            'Dolores',
-            'El Espinal',
-            'Falán',
-            'Flandes',
-            'Fresno',
-            'Guamo',
-            'Herveo',
-            'Honda',
-            'Ibagué',
-            'Icononzo',
-            'Lérida',
-            'Líbano',
-            'Mariquita',
-            'Melgar',
-            'Murillo',
-            'Natagaima',
-            'Ortega',
-            'Palocabildo',
-            'Piedras',
-            'Planadas',
-            'Prado',
-            'Purificación',
-            'Rioblanco',
-            'Roncesvalles',
-            'Rovira',
-            'Saldaña',
-            'San Antonio',
-            'San Luis',
-            'Santa Isabel',
-            'Suárez',
-            'Valle de San Juan',
-            'Venadillo',
-            'Villahermosa',
-            'Villarrica'
-        ]
-    },
-    {
-        department: 'Valle del Cauca',
-        cities: [
-            'Alcalá',
-            'Andalucía',
-            'Ansermanuevo',
-            'Argelia',
-            'Bolívar',
-            'Buenaventura',
-            'Buga',
-            'Bugalagrande',
-            'Caicedonia',
-            'Cali',
-            'Calima',
-            'Candelaria',
-            'Cartago',
-            'Dagua',
-            'El Águila',
-            'El Cairo',
-            'El Cerrito',
-            'El Dovio',
-            'Florida',
-            'Ginebra',
-            'Guacarí',
-            'Jamundí',
-            'La Cumbre',
-            'La Unión',
-            'La Victoria',
-            'Obando',
-            'Palmira',
-            'Pradera',
-            'Restrepo',
-            'Riofrío',
-            'Roldanillo',
-            'San Pedro',
-            'Sevilla',
-            'Toro',
-            'Trujillo',
-            'Tuluá',
-            'Ulloa',
-            'Versalles',
-            'Vijes',
-            'Yotoco',
-            'Yumbo',
-            'Zarzal'
-        ]
-    },
-    {
-        department: 'Vaupés',
-        cities: [
-            'Carurú',
-            'Mitú',
-            'Taraira'
-        ]
-    },
-    {
-        department: 'Vichada',
-        cities: [
-            'Cumaribo',
-            'La Primavera',
-            'Puerto Carreño',
-            'Santa Rosalía'
-        ]
-    },
-    {
-        department: 'Consulados',
-        cities: [
-            'Argel',
-            'El Cairo',
-            'Adís Abeba',
-            'Acra',
-            'Nairobi',
-            'Rabat',
-            'Pretoria',
-            'Buenos Aires',
-            'Córdoba',
-            'La Paz',
-            'Brasilia',
-            'Manaos',
-            'Sao Paulo',
-            'Tabatinga',
-            'Calgary',
-            'Montreal',
-            'Ottawa',
-            'Toronto',
-            'Vancouver',
-            'Antofagasta',
-            'Santiago de Chile',
-            'San José',
-            'La Habana',
-            'Esmeraldas',
-            'Guayaquil',
-            'Nueva Loja',
-            'Quito',
-            'Santo Domingo de los Colorados',
-            'Tulcán',
-            'San Salvador',
-            'Atlanta',
-            'Boston',
-            'Chicago',
-            'Houston',
-            'Los Ángeles',
-            'Miami',
-            'Newark',
-            'Nueva York',
-            'Orlando',
-            'San Francisco',
-            'San Juan, Puerto Rico',
-            'Washington D.C.',
-            'Ciudad de Guatemala',
-            'Tegucigalpa',
-            'Kingston',
-            'Cancún',
-            'Ciudad de México',
-            'Guadalajara',
-            'Managua',
-            'Colón',
-            'Jaqué',
-            'Panamá',
-            'Puerto Obaldía',
-            'Asunción',
-            'Iquitos',
-            'Lima',
-            'Santo Domingo',
-            'Puerto España',
-            'Montevideo',
-            'Barinas',
-            'Barquisimeto',
-            'Caracas',
-            'El Amparo',
-            'Machiques',
-            'Maracaibo',
-            'Mérida',
-            'Puerto Ayacucho',
-            'Puerto La Cruz',
-            'Puerto Ordaz',
-            'San Antonio del Táchira',
-            'San Carlos del Zulia',
-            'San Cristóbal',
-            'San Fernando de Atabapo',
-            'Valencia',
-            'Bakú',
-            'Cantón',
-            'Hong Kong',
-            'Pekín',
-            'Shanghái',
-            'Seúl',
-            'Abu Dabi',
-            'Manila',
-            'Nueva Delhi',
-            'Yakarta',
-            'Tel Aviv',
-            'Tokio',
-            'Beirut',
-            'Kuala Lumpur',
-            'Singapur',
-            'Bangkok',
-            'Ankara',
-            'Hanoi',
-            'Berlín',
-            'Fráncfort',
-            'Viena',
-            'Bruselas',
-            'Barcelona',
-            'Bilbao',
-            'Las Palmas de Gran Canaria',
-            'Madrid',
-            'Palma de Mallorca',
-            'Sevilla',
-            'Valencia',
-            'Helsinki',
-            'París',
-            'Budapest',
-            'Dublín',
-            'Milán',
-            'Roma',
-            'Oslo',
-            'Ámsterdam',
-            'La Haya',
-            'Oranjestad, Aruba',
-            'Willemstad, Antillas Neerlandesas',
-            'Varsovia',
-            'Lisboa',
-            'Londres',
-            'Bucarest',
-            'Moscú',
-            'Roma',
-            'Berna',
-            'Estocolmo',
-            'Kiev',
-            'Canberra',
-            'Sídney',
-            'Auckland'
-        ]
-    }
-];
-
-
-/***/ })
-
-}]);
-//# sourceMappingURL=default~detalles-miembro-detalles-miembro-module~registro-registro-module~views-mi-cuenta-mi-cuenta-module.js.map
