+++ www/main.js
@@ -371,6 +371,23 @@ let GlobalService = class GlobalService {
     getCurrency(amount = 0) {
         return new Intl.NumberFormat('es-CO', { style: 'currency', currency: 'COP' }).format(amount);
     }
+    downloadCSVFromJson(filename, arrayOfJson) {
+        // convert JSON to CSV
+        const replacer = (key, value) => value === null ? '' : value; // specify how you want to handle null values here
+        const header = Object.keys(arrayOfJson[0]);
+        let csv = arrayOfJson.map(row => header.map(fieldName => JSON.stringify(row[fieldName], replacer)).join(','));
+        csv.unshift(header.join(','));
+        csv = csv.join('\r\n');
+        // Create link and download
+        const link = document.createElement('a');
+        link.setAttribute('href', 'data:text/csv;charset=utf-8,%EF%BB%BF' + encodeURIComponent(csv));
+        link.setAttribute('download', filename);
+        link.style.visibility = 'hidden';
+        document.body.appendChild(link);
+        link.click();
+        document.body.removeChild(link);
+    }
+    ;
 };
 GlobalService.ctorParameters = () => [
     { type: _ionic_angular__WEBPACK_IMPORTED_MODULE_2__["AlertController"] },
@@ -962,7 +979,7 @@ const routes = [
     // },
     {
         path: 'miembros',
-        loadChildren: () => Promise.all(/*! import() | views-miembros-miembros-module */[__webpack_require__.e("default~views-ajustes-ajustes-module~views-consolidados-consolidados-module~views-cuentas-bancarias-~291b62ff"), __webpack_require__.e("common"), __webpack_require__.e("views-miembros-miembros-module")]).then(__webpack_require__.bind(null, /*! ./views/miembros/miembros.module */ "PI0l")).then(m => m.MiembrosPageModule)
+        loadChildren: () => Promise.all(/*! import() | views-miembros-miembros-module */[__webpack_require__.e("default~views-ajustes-ajustes-module~views-consolidados-consolidados-module~views-cuentas-bancarias-~291b62ff"), __webpack_require__.e("default~views-mi-cuenta-mi-cuenta-module~views-miembros-miembros-module"), __webpack_require__.e("common"), __webpack_require__.e("views-miembros-miembros-module")]).then(__webpack_require__.bind(null, /*! ./views/miembros/miembros.module */ "PI0l")).then(m => m.MiembrosPageModule)
     },
     {
         path: 'ingresar',
@@ -974,7 +991,7 @@ const routes = [
     },
     {
         path: 'mi-cuenta',
-        loadChildren: () => Promise.all(/*! import() | views-mi-cuenta-mi-cuenta-module */[__webpack_require__.e("default~views-ajustes-ajustes-module~views-consolidados-consolidados-module~views-cuentas-bancarias-~291b62ff"), __webpack_require__.e("default~detalles-miembro-detalles-miembro-module~registro-registro-module~views-mi-cuenta-mi-cuenta-module"), __webpack_require__.e("views-mi-cuenta-mi-cuenta-module")]).then(__webpack_require__.bind(null, /*! ./views/mi-cuenta/mi-cuenta.module */ "1LWO")).then(m => m.MiCuentaPageModule)
+        loadChildren: () => Promise.all(/*! import() | views-mi-cuenta-mi-cuenta-module */[__webpack_require__.e("default~views-ajustes-ajustes-module~views-consolidados-consolidados-module~views-cuentas-bancarias-~291b62ff"), __webpack_require__.e("default~views-mi-cuenta-mi-cuenta-module~views-miembros-miembros-module"), __webpack_require__.e("views-mi-cuenta-mi-cuenta-module")]).then(__webpack_require__.bind(null, /*! ./views/mi-cuenta/mi-cuenta.module */ "1LWO")).then(m => m.MiCuentaPageModule)
     },
     {
         path: 'grupos-familiares',
