+++ www/views-consolidados-consolidados-module.js
@@ -168,6 +168,7 @@ let ConsolidadosPage = class ConsolidadosPage {
             const data = yield this.consolidadosService.getConsolidatesReport(this.queryParams);
             if (data && !data.error) {
                 this.reports = data;
+                console.log('data', data);
                 this.members = data.members || [];
                 this.setConsolidatesValues(data.consolidates || []);
                 this.setPendingMembers(data.pendingVisits || []);
@@ -291,6 +292,31 @@ let ConsolidadosPage = class ConsolidadosPage {
             yield this.router.navigate([`consolidados/registrar-visita`]);
         });
     }
+    downloadData() {
+        if (this.consolidates.length > 0) {
+            const model = [];
+            // const keys = [
+            //   { key: 'date', key2: 'Fecha' },
+            //   { key: 'consolidator', key2: 'Visitante' },
+            //   { key: 'member', key2: 'Visitado' },
+            //   { key: 'observation', key2: 'Observacion' },
+            // ];
+            const getNames = (member = null) => {
+                return member ? `${member.names} ${member.lastNames} - ${member.phone}` : null;
+            };
+            this.consolidates.forEach(c => {
+                var _a;
+                const m = {
+                    Fecha: c.date,
+                    Visitador: getNames(c.consolidator || null),
+                    Visitado: getNames(c.member || null),
+                    Observacion: (_a = c.observation) === null || _a === void 0 ? void 0 : _a.replaceAll(',', ';;'),
+                };
+                model.push(m);
+            });
+            this.globalSer.downloadCSVFromJson(`VISITAS_CONSOLIDADOS_CCADV_${dayjs__WEBPACK_IMPORTED_MODULE_5___default()().format('YYYY-MM-DD[_]HH:mm:ss')}.csv`, model);
+        }
+    }
 };
 ConsolidadosPage.ctorParameters = () => [
     { type: _services_global_service__WEBPACK_IMPORTED_MODULE_11__["GlobalService"] },
@@ -331,7 +357,7 @@ __webpack_require__.r(__webpack_exports__);
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony default export */ __webpack_exports__["default"] = ("<ion-header>\n  <ion-toolbar>\n    <ion-title>Consolidados</ion-title>\n    <ion-buttons slot=\"end\">\n      <ion-button (click)=\"showListResizeAlert()\">\n        <ion-icon name=\"resize-outline\" slot=\"start\"></ion-icon>\n        <ion-text>Visualización</ion-text>\n      </ion-button>\n    </ion-buttons>\n  </ion-toolbar>\n</ion-header>\n\n<ion-content>\n  <ion-row class=\"ion-no-padding ion-margin-start ion-margin-end ion-margin-top\">\n    <ion-col size=\"12\">\n      <div class=\"\">\n        <ion-button color=\"medium\" (click)=\"setView('filter')\">\n          <ion-icon name=\"filter-outline\" slot=\"start\" *ngIf=\"!views.filter\"></ion-icon>\n          <ion-icon name=\"close-outline\" slot=\"start\" *ngIf=\"views.filter\"></ion-icon>\n          <ion-text>Filtrar</ion-text>\n        </ion-button>\n        <ion-button (click)=\"goToVisitRegisterForm()\">\n          <ion-icon name=\"clipboard-outline\" slot=\"start\"></ion-icon>\n          <ion-text>Registrar visita</ion-text>\n        </ion-button>\n      </div>\n    </ion-col>\n    <ion-col size=\"12\" *ngIf=\"views.filter\">\n      <ion-card>\n        <ion-card-content>\n          <ion-row>\n            <ion-col size=\"12\" size-sm=\"6\">\n              <ion-item>\n                <ion-label position=\"floating\" color=\"medium\">A partir de:</ion-label>\n                <ion-datetime\n                  [(ngModel)]=\"queryParams.initDate\"\n                  displayFormat=\"DD/MM/YYYY\"\n                  [min]=\"minInitDate\"\n                  [max]=\"maxInitDate\"\n                  doneText=\"Listo\"\n                  cancelText=\"Cancelar\"\n                ></ion-datetime>\n              </ion-item>\n            </ion-col>\n            <ion-col size=\"12\" size-sm=\"6\">\n              <ion-item>\n                <ion-label position=\"floating\" color=\"medium\">Hasta:</ion-label>\n                <ion-datetime\n                  [(ngModel)]=\"queryParams.endDate\"\n                  displayFormat=\"DD/MM/YYYY\"\n                  [min]=\"minEndDate\"\n                  [max]=\"maxEndDate\"\n                  doneText=\"Listo\"\n                  cancelText=\"Cancelar\"\n                ></ion-datetime>\n              </ion-item>\n            </ion-col>\n          </ion-row>\n          <ion-row class=\"ion-margin-top\">\n            <ion-col size=\"6\" class=\"ion-justify-content-center\">\n              <ion-button class=\"ion-text-center\" color=\"medium\" expand=\"full\" (click)=\"resetQueryParams()\">\n                <ion-icon name=\"trash-outline\"></ion-icon>\n                <ion-text>Limpiar</ion-text>\n              </ion-button>\n            </ion-col>\n            <ion-col size=\"6\" class=\"ion-justify-content-center\">\n              <ion-button class=\"ion-text-center\" color=\"primary\" expand=\"full\" (click)=\"findData()\">\n                <ion-icon name=\"checkmark-outline\"></ion-icon>\n                <ion-text>Aplicar</ion-text>\n              </ion-button>\n            </ion-col>\n          </ion-row>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n  </ion-row>\n\n  <ion-row class=\"ion-margin-top ion-margin-bottom ion-padding-bottom\" *ngIf=\"reports\">\n    <ion-col size=\"12\">\n      <ion-card class=\"ion-no-padding\">\n        <ion-card-header class=\"ion-no-padding\">\n          <ion-toolbar color=\"primary\">\n            <ion-buttons slot=\"start\">\n              <ion-button (click)=\"setView('totals')\">\n                <ion-icon name=\"chevron-{{views.totals  ? 'up' : 'down' }}-outline\"></ion-icon>\n              </ion-button>\n            </ion-buttons>\n            <ion-title>Resumen</ion-title>\n          </ion-toolbar>\n        </ion-card-header>\n        <ion-card-content class=\"{{ views.totals ? '' : 'ion-hide'}} ion-no-padding no-selected-text\">\n          <ion-row class=\"ion-margin-top ion-margin-bottom\">\n            <ion-col size=\"12\" [sizeSm]=\"4\" class=\"ion-text-center\">\n              <ion-text>\n                <span class=\"bigIcon\">{{ members.length }}</span>\n                <br>\n              </ion-text>\n              <ion-text>\n                <p class=\"textTotals\">\n                  <ion-icon name=\"body-outline\" class=\"textTotals\"></ion-icon>\n                  Miembros consolidados\n                </p>\n              </ion-text>\n            </ion-col>\n            <ion-col size=\"12\" [sizeSm]=\"4\" class=\"ion-text-center\">\n              <ion-text>\n                <span class=\"bigIcon\">{{ consolidates.length }}</span>\n                <br>\n              </ion-text>\n              <ion-text>\n                <p class=\"textTotals\">\n                  <ion-icon name=\"footsteps-outline\" class=\"textTotals\"></ion-icon>\n                  Visitas realizadas\n                </p>\n              </ion-text>\n            </ion-col>\n            <ion-col size=\"12\" [sizeSm]=\"4\" class=\"ion-text-center\">\n              <ion-text>\n                <span class=\"bigIcon\">{{ pendingVisitsTotals || 0 }}</span>\n                <br>\n              </ion-text>\n              <ion-text>\n                <p class=\"textTotals\">\n                  <ion-icon name=\"timer-outline\" class=\"textTotals\"></ion-icon>\n                  Visitas pendientes\n                </p>\n              </ion-text>\n            </ion-col>\n          </ion-row>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n    <ion-col size=\"12\" [sizeSm]=\"sizeElems\">\n      <ion-card class=\"ion-no-padding\">\n        <ion-card-header class=\"ion-no-padding\">\n          <ion-toolbar color=\"primary\">\n            <ion-buttons slot=\"start\">\n              <ion-button (click)=\"setView('visits')\">\n                <ion-icon name=\"chevron-{{views.visits ? 'up' : 'down' }}-outline\"></ion-icon>\n              </ion-button>\n            </ion-buttons>\n            <ion-title>Visitas</ion-title>\n          </ion-toolbar>\n        </ion-card-header>\n        <ion-card-content class=\"{{ views.visits ? '' : 'ion-hide'}} ion-no-padding\">\n          <ion-row *ngIf=\"consolidatesPreview && consolidatesPreview.length > 0; else noConsolidates\" class=\"{{ sizeElems === 6 ? 'max-height-250' : 'max-height-450'}}\">\n            <ion-col size=\"12\" [sizeSm]=\"sizeElems === 6 ? 12 : 6\" *ngFor=\"let c of consolidatesPreview; let i=index\">\n              <ion-card (click)=\"showDetails(i)\">\n                <ion-card-content class=\"no-selected-text\">\n                  <ion-text>\n                    <p><b>{{ c.member ? c.member.names + ' ' + c.member.lastNames : 'Desconocido' }}</b></p>\n                  </ion-text>\n                  <ion-text>\n                    <p>{{ c.observation }}</p>\n                  </ion-text>\n                  <ion-text>\n                    <span><b>Fecha: </b>{{ c.date }}</span><br>\n                  </ion-text>\n                  <ion-text>\n                    <i class=\"toSmall\">Clic para más detalles.</i>\n                  </ion-text>\n                </ion-card-content>\n              </ion-card>\n            </ion-col>\n          </ion-row>\n          <ng-template #noConsolidates>\n            <ion-row class=\"ion-margin-top ion-margin-bottom\">\n              <ion-col size=\"12\">\n                <ion-text color=\"medium\">\n                  <p class=\"ion-text-center\">\n                    <ion-icon name=\"alert-circle-outline\"></ion-icon> No se encontraron visitas.\n                  </p>\n                </ion-text>\n              </ion-col>\n            </ion-row>\n          </ng-template>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n    <ion-col size=\"12\" [sizeSm]=\"sizeElems\">\n      <ion-card class=\"ion-no-padding\">\n        <ion-card-header class=\"ion-no-padding\">\n          <ion-toolbar color=\"primary\">\n            <ion-buttons slot=\"start\">\n              <ion-button (click)=\"setView('members')\">\n                <ion-icon name=\"chevron-{{views.members ? 'up' : 'down' }}-outline\"></ion-icon>\n              </ion-button>\n            </ion-buttons>\n            <ion-title>Miembros consolidados</ion-title>\n          </ion-toolbar>\n        </ion-card-header>\n        <ion-card-content class=\"{{ views.members ? '' : 'ion-hide'}} ion-no-padding\">\n          <ion-row *ngIf=\"members && members.length > 0; else noMembers\" class=\"{{ sizeElems === 6 ? 'max-height-250' : 'max-height-450'}}\">\n            <ion-col size=\"12\" [sizeSm]=\"sizeElems === 6 ? 12 : 4\" *ngFor=\"let member of members;\">\n              <app-usercard [data]=\"member\"></app-usercard>\n            </ion-col>\n          </ion-row>\n          <ng-template #noMembers>\n            <ion-row class=\"ion-margin-top ion-margin-bottom\">\n              <ion-col size=\"12\">\n                <ion-text color=\"medium\">\n                  <p class=\"ion-text-center\">\n                    <ion-icon name=\"alert-circle-outline\"></ion-icon> No existen miembros.\n                  </p>\n                </ion-text>\n              </ion-col>\n            </ion-row>\n          </ng-template>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n    <ion-col size=\"12\" [sizeSm]=\"sizeElems\">\n      <ion-card class=\"ion-no-padding\">\n        <ion-card-header class=\"ion-no-padding\">\n          <ion-toolbar color=\"primary\">\n            <ion-buttons slot=\"start\">\n              <ion-button (click)=\"setView('pending')\">\n                <ion-icon name=\"chevron-{{views.pending ? 'up' : 'down' }}-outline\"></ion-icon>\n              </ion-button>\n            </ion-buttons>\n            <ion-title>Miembros pendientes por visita</ion-title>\n          </ion-toolbar>\n        </ion-card-header>\n        <ion-card-content class=\"{{ views.pending ? '' : 'ion-hide'}} ion-no-padding\">\n          <ion-row *ngIf=\"pendingMembers && pendingMembers.length > 0; else noPending\" class=\"{{ sizeElems === 6 ? 'max-height-250' : 'max-height-450'}}\">\n            <ion-col size=\"12\" [sizeSm]=\"sizeElems === 6 ? 12 : 4\" *ngFor=\"let pm of pendingMembers;\">\n              <app-usercard [data]=\"pm\" [goDetails]=\"false\"></app-usercard>\n            </ion-col>\n          </ion-row>\n          <ng-template #noPending>\n            <ion-row class=\"ion-margin-top ion-margin-bottom\">\n              <ion-col size=\"12\">\n                <ion-text color=\"medium\">\n                  <p class=\"ion-text-center\">\n                    <ion-icon name=\"alert-circle-outline\"></ion-icon> No existen miembros pendientes por visitas.\n                  </p>\n                </ion-text>\n              </ion-col>\n            </ion-row>\n          </ng-template>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n  </ion-row>\n\n</ion-content>\n");
+/* harmony default export */ __webpack_exports__["default"] = ("<ion-header>\n  <ion-toolbar>\n    <ion-title>Consolidados</ion-title>\n    <ion-buttons slot=\"end\">\n      <ion-button (click)=\"showListResizeAlert()\">\n        <ion-icon name=\"resize-outline\" slot=\"start\"></ion-icon>\n        <ion-text>Visualización</ion-text>\n      </ion-button>\n    </ion-buttons>\n  </ion-toolbar>\n</ion-header>\n\n<ion-content>\n  <ion-row class=\"ion-no-padding ion-margin-start ion-margin-end ion-margin-top\">\n    <ion-col size=\"12\">\n      <div class=\"\">\n        <ion-button color=\"medium\" (click)=\"setView('filter')\">\n          <ion-icon name=\"filter-outline\" slot=\"start\" *ngIf=\"!views.filter\"></ion-icon>\n          <ion-icon name=\"close-outline\" slot=\"start\" *ngIf=\"views.filter\"></ion-icon>\n          <ion-text>Filtrar</ion-text>\n        </ion-button>\n        <ion-button (click)=\"goToVisitRegisterForm()\">\n          <ion-icon name=\"clipboard-outline\" slot=\"start\"></ion-icon>\n          <ion-text>Registrar visita</ion-text>\n        </ion-button>\n        <ion-button (click)=\"downloadData()\">\n          <ion-icon name=\"download-outline\" slot=\"start\"></ion-icon>\n          <ion-text>Descargar datos</ion-text>\n        </ion-button>\n      </div>\n    </ion-col>\n    <ion-col size=\"12\" *ngIf=\"views.filter\">\n      <ion-card>\n        <ion-card-content>\n          <ion-row>\n            <ion-col size=\"12\" size-sm=\"6\">\n              <ion-item>\n                <ion-label position=\"floating\" color=\"medium\">A partir de:</ion-label>\n                <ion-datetime\n                  [(ngModel)]=\"queryParams.initDate\"\n                  displayFormat=\"DD/MM/YYYY\"\n                  [min]=\"minInitDate\"\n                  [max]=\"maxInitDate\"\n                  doneText=\"Listo\"\n                  cancelText=\"Cancelar\"\n                ></ion-datetime>\n              </ion-item>\n            </ion-col>\n            <ion-col size=\"12\" size-sm=\"6\">\n              <ion-item>\n                <ion-label position=\"floating\" color=\"medium\">Hasta:</ion-label>\n                <ion-datetime\n                  [(ngModel)]=\"queryParams.endDate\"\n                  displayFormat=\"DD/MM/YYYY\"\n                  [min]=\"minEndDate\"\n                  [max]=\"maxEndDate\"\n                  doneText=\"Listo\"\n                  cancelText=\"Cancelar\"\n                ></ion-datetime>\n              </ion-item>\n            </ion-col>\n          </ion-row>\n          <ion-row class=\"ion-margin-top\">\n            <ion-col size=\"6\" class=\"ion-justify-content-center\">\n              <ion-button class=\"ion-text-center\" color=\"medium\" expand=\"full\" (click)=\"resetQueryParams()\">\n                <ion-icon name=\"trash-outline\"></ion-icon>\n                <ion-text>Limpiar</ion-text>\n              </ion-button>\n            </ion-col>\n            <ion-col size=\"6\" class=\"ion-justify-content-center\">\n              <ion-button class=\"ion-text-center\" color=\"primary\" expand=\"full\" (click)=\"findData()\">\n                <ion-icon name=\"checkmark-outline\"></ion-icon>\n                <ion-text>Aplicar</ion-text>\n              </ion-button>\n            </ion-col>\n          </ion-row>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n  </ion-row>\n\n  <ion-row class=\"ion-margin-top ion-margin-bottom ion-padding-bottom\" *ngIf=\"reports\">\n    <ion-col size=\"12\">\n      <ion-card class=\"ion-no-padding\">\n        <ion-card-header class=\"ion-no-padding\">\n          <ion-toolbar color=\"primary\">\n            <ion-buttons slot=\"start\">\n              <ion-button (click)=\"setView('totals')\">\n                <ion-icon name=\"chevron-{{views.totals  ? 'up' : 'down' }}-outline\"></ion-icon>\n              </ion-button>\n            </ion-buttons>\n            <ion-title>Resumen</ion-title>\n          </ion-toolbar>\n        </ion-card-header>\n        <ion-card-content class=\"{{ views.totals ? '' : 'ion-hide'}} ion-no-padding no-selected-text\">\n          <ion-row class=\"ion-margin-top ion-margin-bottom\">\n            <ion-col size=\"12\" [sizeSm]=\"4\" class=\"ion-text-center\">\n              <ion-text>\n                <span class=\"bigIcon\">{{ members.length }}</span>\n                <br>\n              </ion-text>\n              <ion-text>\n                <p class=\"textTotals\">\n                  <ion-icon name=\"body-outline\" class=\"textTotals\"></ion-icon>\n                  Miembros consolidados\n                </p>\n              </ion-text>\n            </ion-col>\n            <ion-col size=\"12\" [sizeSm]=\"4\" class=\"ion-text-center\">\n              <ion-text>\n                <span class=\"bigIcon\">{{ consolidates.length }}</span>\n                <br>\n              </ion-text>\n              <ion-text>\n                <p class=\"textTotals\">\n                  <ion-icon name=\"footsteps-outline\" class=\"textTotals\"></ion-icon>\n                  Visitas realizadas\n                </p>\n              </ion-text>\n            </ion-col>\n            <ion-col size=\"12\" [sizeSm]=\"4\" class=\"ion-text-center\">\n              <ion-text>\n                <span class=\"bigIcon\">{{ pendingVisitsTotals || 0 }}</span>\n                <br>\n              </ion-text>\n              <ion-text>\n                <p class=\"textTotals\">\n                  <ion-icon name=\"timer-outline\" class=\"textTotals\"></ion-icon>\n                  Visitas pendientes\n                </p>\n              </ion-text>\n            </ion-col>\n          </ion-row>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n    <ion-col size=\"12\" [sizeSm]=\"sizeElems\">\n      <ion-card class=\"ion-no-padding\">\n        <ion-card-header class=\"ion-no-padding\">\n          <ion-toolbar color=\"primary\">\n            <ion-buttons slot=\"start\">\n              <ion-button (click)=\"setView('visits')\">\n                <ion-icon name=\"chevron-{{views.visits ? 'up' : 'down' }}-outline\"></ion-icon>\n              </ion-button>\n            </ion-buttons>\n            <ion-title>Visitas</ion-title>\n          </ion-toolbar>\n        </ion-card-header>\n        <ion-card-content class=\"{{ views.visits ? '' : 'ion-hide'}} ion-no-padding\">\n          <ion-row *ngIf=\"consolidatesPreview && consolidatesPreview.length > 0; else noConsolidates\" class=\"{{ sizeElems === 6 ? 'max-height-250' : 'max-height-450'}}\">\n            <ion-col size=\"12\" [sizeSm]=\"sizeElems === 6 ? 12 : 6\" *ngFor=\"let c of consolidatesPreview; let i=index\">\n              <ion-card (click)=\"showDetails(i)\">\n                <ion-card-content class=\"no-selected-text\">\n                  <ion-text>\n                    <p><b>{{ c.member ? c.member.names + ' ' + c.member.lastNames : 'Desconocido' }}</b></p>\n                  </ion-text>\n                  <ion-text>\n                    <p>{{ c.observation }}</p>\n                  </ion-text>\n                  <ion-text>\n                    <span><b>Fecha: </b>{{ c.date }}</span><br>\n                  </ion-text>\n                  <ion-text>\n                    <i class=\"toSmall\">Clic para más detalles.</i>\n                  </ion-text>\n                </ion-card-content>\n              </ion-card>\n            </ion-col>\n          </ion-row>\n          <ng-template #noConsolidates>\n            <ion-row class=\"ion-margin-top ion-margin-bottom\">\n              <ion-col size=\"12\">\n                <ion-text color=\"medium\">\n                  <p class=\"ion-text-center\">\n                    <ion-icon name=\"alert-circle-outline\"></ion-icon> No se encontraron visitas.\n                  </p>\n                </ion-text>\n              </ion-col>\n            </ion-row>\n          </ng-template>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n    <ion-col size=\"12\" [sizeSm]=\"sizeElems\">\n      <ion-card class=\"ion-no-padding\">\n        <ion-card-header class=\"ion-no-padding\">\n          <ion-toolbar color=\"primary\">\n            <ion-buttons slot=\"start\">\n              <ion-button (click)=\"setView('members')\">\n                <ion-icon name=\"chevron-{{views.members ? 'up' : 'down' }}-outline\"></ion-icon>\n              </ion-button>\n            </ion-buttons>\n            <ion-title>Miembros consolidados</ion-title>\n          </ion-toolbar>\n        </ion-card-header>\n        <ion-card-content class=\"{{ views.members ? '' : 'ion-hide'}} ion-no-padding\">\n          <ion-row *ngIf=\"members && members.length > 0; else noMembers\" class=\"{{ sizeElems === 6 ? 'max-height-250' : 'max-height-450'}}\">\n            <ion-col size=\"12\" [sizeSm]=\"sizeElems === 6 ? 12 : 4\" *ngFor=\"let member of members;\">\n              <app-usercard [data]=\"member\"></app-usercard>\n            </ion-col>\n          </ion-row>\n          <ng-template #noMembers>\n            <ion-row class=\"ion-margin-top ion-margin-bottom\">\n              <ion-col size=\"12\">\n                <ion-text color=\"medium\">\n                  <p class=\"ion-text-center\">\n                    <ion-icon name=\"alert-circle-outline\"></ion-icon> No existen miembros.\n                  </p>\n                </ion-text>\n              </ion-col>\n            </ion-row>\n          </ng-template>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n    <ion-col size=\"12\" [sizeSm]=\"sizeElems\">\n      <ion-card class=\"ion-no-padding\">\n        <ion-card-header class=\"ion-no-padding\">\n          <ion-toolbar color=\"primary\">\n            <ion-buttons slot=\"start\">\n              <ion-button (click)=\"setView('pending')\">\n                <ion-icon name=\"chevron-{{views.pending ? 'up' : 'down' }}-outline\"></ion-icon>\n              </ion-button>\n            </ion-buttons>\n            <ion-title>Miembros pendientes por visita</ion-title>\n          </ion-toolbar>\n        </ion-card-header>\n        <ion-card-content class=\"{{ views.pending ? '' : 'ion-hide'}} ion-no-padding\">\n          <ion-row *ngIf=\"pendingMembers && pendingMembers.length > 0; else noPending\" class=\"{{ sizeElems === 6 ? 'max-height-250' : 'max-height-450'}}\">\n            <ion-col size=\"12\" [sizeSm]=\"sizeElems === 6 ? 12 : 4\" *ngFor=\"let pm of pendingMembers;\">\n              <app-usercard [data]=\"pm\" [goDetails]=\"false\"></app-usercard>\n            </ion-col>\n          </ion-row>\n          <ng-template #noPending>\n            <ion-row class=\"ion-margin-top ion-margin-bottom\">\n              <ion-col size=\"12\">\n                <ion-text color=\"medium\">\n                  <p class=\"ion-text-center\">\n                    <ion-icon name=\"alert-circle-outline\"></ion-icon> No existen miembros pendientes por visitas.\n                  </p>\n                </ion-text>\n              </ion-col>\n            </ion-row>\n          </ng-template>\n        </ion-card-content>\n      </ion-card>\n    </ion-col>\n  </ion-row>\n\n</ion-content>\n");
 
 /***/ }),
